Implemented basic non-blocking TCP/IP sockets.
Now client and server can both have some interaction interaction:
the server sends a welcome message to the client, and prints
everything the client sends it. The next step would be to implement
a basic CLI interface for interacting with the server.

I also changed the code style to K&R (kind of).